requirements.txt
langchain
langchain-huggingface
huggingface-hub
python-hcl2
transformers
torch
accelerate